add/get API calls tested; TODO debug leave and delete calls

+++ controllers/api.py
@@ -1,22 +1,21 @@
-# database retrieve information for client ------
-
+# database retrieve information for client ------ -------------------------------------------------
+@auth.requires_login()
 def get_classes():
     """get the classes associated with a student"""
-    # select all relations between students and classes for particular student
-    student = get_student(request.vars.auth_id)
-    if student is None:
-        return response.json(dict())
-    classes = db(db.class_students.student_id == student).select()
+    # select all relations between user and classes for particular student
+    class_refs = db(db.class_users.user_ref == auth.user).select()
     response_classes = []
+
     # get class from each relation (classes of student)
-    for i, cref in enumerate(classes):
-        response_classes.append(response_class(cref.class_id))
+    for i, cref in enumerate(class_refs):
+        response_classes.append(response_class(cref.class_ref))
     return response.json(dict(classes=response_classes))
 
 
+@auth.requires_login()
 def get_projects():
     """get all projects associated with a class"""
-    projects = db(db.projects.class_id == request.vars.class_id).select()
+    projects = db(db.projects.class_ref == request.vars.class_id).select()
     response_projects = []
 
     for i, p in enumerate(projects):
@@ -24,74 +23,94 @@ def get_projects():
     return response.json(dict(projects=response_projects))
 
 
+@auth.requires_login()
 def get_groups():
     """get the groups associated with a class"""
     # select all groups for a class
-    groups = db(db.groups.class_id == request.vars.class_id).select()
+    groups = db(db.groups.project_ref == request.vars.project_id).select()
     response_groups = []
+
     # groups are only associated to one class per
     for i, g in enumerate(groups):
         response_groups.append(response_group(g))
     return response.json(dict(groups=response_groups))
 
 
+@auth.requires_login()
 def get_members():
     """get the students associated with a group"""
     # select all relations between groups and students for particular group
-    students = db(db.group_students.group_id == request.vars.group_id).select()
+    member_refs = db(db.group_students.group_ref == request.vars.group_id).select()
     response_members = []
+
     # get student from each relation (members of group)
-    for i, sref in enumerate(students):
-        response_members.append(response_member(sref.student_id))
+    for i, mref in enumerate(member_refs):
+        response_members.append(response_member(mref.student_ref))
     return response.json(dict(members=response_members))
 
 
-# database insert and delete --------------------
-
-def add_student():
-    """add a new student to the database"""
-    auth_id = request.vars.auth_id
-    if db.students(db.students.google_auth_id==auth_id) is None:
-        db.students.insert(google_auth_id=auth_id)
-        return response.json('add student success')
-    return response.json('already added student')
+# instructor permissions ------------------------
+def auth_instructor_class(class_id):
+    if auth.user.id != db.classes[class_id].instructor_ref.id:
+        raise HTTP(401)
 
 
-def delete_student():
-    """delete a student from the database"""
-    auth_id = request.vars.auth_id
-    db(db.students.google_auth_id == auth_id).delete()
-    return response.json('delete student success')
+def auth_instructor_project(project_id):
+    if auth.user.id != db.projects[project_id].class_ref.instructor_ref.id:
+        raise HTTP(401)
 
 
+def auth_instructor_group(group_id):
+    if auth.user.id == db.groups[group_id].project_ref.class_ref.instructor_ref.id:
+        return True
+    return False
+
+
+# leader permissions ----------------------------
+def auth_leader_group(group_id):
+    if auth.user.id is not db.groups[group_id].leader_ref.id:
+        return True
+    return False
+
+
+# database insert and delete -------------------- -------------------------------------------------
 # class api calls -------------------------------
+@auth.requires_login()
 def create_class():
     """a student creates a class and is the 'instructor'"""
     name = request.vars.name
     description = request.vars.description
-    instructor = get_student(request.vars.auth_id)
+    instructor = auth.user
 
-    class_id = db.classes.insert(name=name, description=description, instructor_id=instructor.id)
-    db.class_students.insert(student_id=instructor.id, class_id=class_id)
+    class_id = db.classes.insert(name=name, description=description, instructor_ref=instructor)
+    db.class_users.insert(user_ref=instructor, class_ref=db.classes[class_id])
     return response.json('create class success')
 
 
+@auth.requires_login()
 def join_class():
     """a student joins a project"""
     class_id = request.vars.class_id
-    student = get_student(request.vars.auth_id)
-    db.class_students.insert(student_id=student, class_id=db.classes[class_id])
-    return response.json('add class success')
+    student = auth.user
+
+    db.class_users.insert(student_ref=student, class_ref=db.classes[class_id])
+    return response.json('join class success')
 
 
+@auth.requires_login()
 def leave_class():
     """a student leaves a class"""
     class_id = request.vars.class_id
-    student = get_student(request.vars.auth_id)
-    db(db.class_students.student_id == student.id and db.class_students.class_id == class_id).delete()
+    student = auth.user
+
+    db(db.class_users.user_ref.id == student.id and db.class_users.class_ref.id == class_id).delete()
+    if db(db.class_users.class_ref.id == class_id).select().first() is None:
+        db(db.classes.id == class_id).delete()
+        return response.json('leave and remove class success')
     return response.json('leave class success')
 
 
+@auth.requires_login()
 def delete_class():
     """delete a class"""
     class_id = request.vars.class_id
@@ -100,100 +119,106 @@ def delete_class():
 
 
 # project api calls -----------------------------
+@auth.requires_login()
 def create_project():
     """only instructor of class may create project"""
     class_id = request.vars.class_id
-    instructor = get_student(request.vars.auth_id)
-    if (auth_creator(request.vars.auth_id, db(db.classes.id == class_id.select().first().instructor_id))):
-        return response.json('invalid; not instructor')
+    auth_instructor_class(class_id)
+
     name = request.vars.name
     description = request.vars.description
 
-    project_id = db.projects.insert(name=name, description=description, class_id=class_id)
-    return response.json('create class success')
+    project_id = db.projects.insert(name=name, description=description, class_ref=db.classes[class_id])
+    return response.json('create project success')
 
 
+@auth.requires_login()
 def join_project():
     """a student joins a project"""
     project_id = request.vars.project_id
-    student = get_student(request.vars.auth_id)
-    db.proj_students.insert(student_id=student.id, project_id=project_id)
-    return response.json('add class success')
+    student = auth.user
+
+    db.proj_students.insert(student_ref=student, project_ref=db.projects[project_id])
+    return response.json('join project success')
 
 
+@auth.requires_login()
 def leave_project():
     """a student leaves a project"""
     project_id = request.vars.project_id
-    student = get_student(request.vars.auth_id)
-    db(db.proj_students.student_id == student.id and db.proj_students.project_id == project_id).delete()
+    student = auth.user
+
+    db(db.proj_students.student_ref == student and db.proj_students.project_ref.id == project_id).delete()
+    if db(db.class_users.project_ref.id == project_id).select().first() is None:
+        db(db.projects.id == project_id).delete()
+        return response.json('leave and remove project success')
     return response.json('leave class success')
 
 
+@auth.requires_login()
 def delete_project():
-    """delete a project"""
+    """instructor deletes a project"""
     project_id = request.vars.project_id
+    auth_instructor_project(project_id)
+
     db(db.projects.id == project_id).delete()
     return response.json('delete class success')
 
 
 # group api calls -------------------------------
+@auth.requires_login()
 def create_group():
     """a student creates a group and is the 'leader'"""
     project_id = request.vars.project_id
     name = request.vars.name
     description = request.vars.description
-    leader = get_student(request.vars.auth_id)
+    leader = auth.user
 
-    group_id = db.groups.insert(name=name, description=description, leader_id=leader.id, project_id=project_id)
-    db.group_students.insert(student_id=leader.id, group_id=group_id)
+    group_id = db.groups.insert(name=name, description=description, leader_ref=leader, project_ref=db.projects[project_id])
+    db.group_students.insert(student_ref=leader, group_ref=db.groups[group_id])
     return response.json('create group success')
 
 
+@auth.requires_login()
 def join_group():
     """a student joins a group"""
     group_id = request.vars.group_id
-    student = get_student(request.vars.auth_id)
-    db.group_students.insert(student_id=student.id, group_id=group_id)
+    student = auth.user
+
+    db.group_students.insert(student_ref=student, group_ref=db.groups[group_id])
     return response.json('join group success')
 
 
+@auth.requires_login()
 def leave_group():
     """a student leaves a group and autodetect 0 members for deletion"""
     group_id = request.vars.group_id
-    student = get_student(request.vars.auth_id)
-    db(db.group_students.group_id == group_id and db.group_students.student_id == student.id).delete()
+    student = auth.user
 
-    print db(db.group_students.group_id == group_id).select()
-    if (db(db.group_students.group_id == group_id).select() is None):
+    db(db.group_students.group_ref.id == group_id and db.group_students.student_ref == student).delete()
+    if (db(db.group_students.group_ref.id == group_id).select().first() is None):
         db(db.groups.id == group_id).delete()
         return response.json('leave and remove group success')
     return response.json('leave group success')
 
 
+@auth.requires_login()
 def delete_group():
     """delete a class"""
     group_id = request.vars.group_id
+    if not auth_instructor_group(group_id) and not auth_leader_group(group_id):
+        raise HTTP(401)
+
     db(db.groups.id == group_id).delete()
     return response.json('delete class success')
 
-
-# check if instructor for class
-def auth_creator(auth_id, creator):
-    student = get_student(auth_id)
-    return student is creator
-
-
+# other methods  -------------------------------- -------------------------------------------------
 # check if the vars exist
 def assert_vars(*vars):
     for var in vars:
         if var is None:
             return False
     return True
-
-
-# get student for google_auth_id
-def get_student(auth_id):
-    return db(db.students.google_auth_id == auth_id).select().first()
 
 
 # helper methods to ensure proper dictionary is sent to client
@@ -226,10 +251,7 @@ def response_group(g):
 
 
 def response_member(s):
-    return dict(
-        auth_id=s.google_auth_id,
-    )
-
+    return s.first_name + ' ' + s.last_name
 
 def pr(msg):
     print(msg)
